requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pandas==2.2.2
python-multipart==0.0.9
aiosqlite==0.20.0
openpyxl==3.1.5
pyjwt==2.12.1
bcrypt==5.0.0
